refactor: create 2 separate referral-discount-notice components for A/B testing
