+++ app/components/pay-page/experimental-referral-discount-notice.hbs
@@ -1,0 +1,31 @@
+<div
+  class="flex p-3.5 bg-gradient-to-b from-gray-600 to-gray-900 border border-yellow-600 rounded relative shadow"
+  data-test-private-leaderboard-feature-suggestion
+  ...attributes
+>
+  <div class="mr-1.5 flex-shrink-0">
+    <img src={{this.iconImage}} alt="treasure" class="w-8 animate-spin-once" />
+  </div>
+
+  <div class="bg-repeating-gradient-to-b from-yellow-100 to-amber-500 text-transparent bg-clip-text leading-7 inline-block mt-0.5">
+    {{#if @discount.affiliateReferral.affiliateLink.overriddenAffiliateAvatarUrl}}
+      <img
+        alt="avatar"
+        src={{@discount.affiliateReferral.affiliateLink.overriddenAffiliateAvatarUrl}}
+        class="inline-block mb-0 w-6 h-6 transform -translate-y-px"
+      />
+    {{else}}
+      <AvatarImage
+        @user={{@discount.affiliateReferral.referrer}}
+        class="inline-block mb-0 w-6 h-6 border border-gray-100 rounded-full transform -translate-y-px"
+      />
+    {{/if}}
+
+    <b>{{@discount.affiliateReferral.affiliateLink.usernameForDisplay}}</b>'s referral offer: Subscribe in
+    <span class="font-bold percy-timestamp font-mono text-lg border-b border-yellow-500 border-dashed">{{this.timeLeftText}}</span>
+    to get
+    <span class="font-bold">40% off</span>
+    the 1 year plan.
+  </div>
+  <EmberTooltip @text="The 1 year plan is usually $360, but you can get it for $216 with this offer." @side="bottom" />
+</div>
